requirements.txt
fractal-defi
pandas
numpy
matplotlib
python-dateutil
requests
mlflow
optuna
